pyproject.toml
[tool.poetry]
name = "m-19-prod"
version = "0.1.0"
description = ""
authors = ["sergei_kolbasin <georgio1997@mail.ru>"]


[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
django-rest-framework = "^0.1.0"
django-filter = "^23.2"
django-debug-toolbar = "^4.1.0"
pillow = "^10.0.0"
gunicorn = "^20.1.0"
sentry-sdk = "^1.28.1"
drf-spectacular = "^0.26.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
